# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Polimoney Hub API",
    "version": "1.0.0",
    "description": "政治資金の透明化を推進する Polimoney プラットフォームの Hub API\n\n## 認証\n\n### API キー認証（公開 API）\n`X-API-Key` ヘッダーに API キーを設定してください。\n\n### Bearer トークン認証（管理者 API）\n`Authorization: Bearer <token>` ヘッダーに Supabase Auth のアクセストークンを設定してください。",
    "contact": {
      "name": "Polimoney Team",
      "url": "https://github.com/moai-redcap/polimoney_hub"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3722",
      "description": "ローカル開発環境"
    },
    {
      "url": "https://polimoney-hub.dd2030-topics.deno.net",
      "description": "本番環境"
    }
  ],
...
  "paths": {
...
    "/api/v1/unlock-requests": {
      "post": {
        "tags": ["Unlock"],
        "summary": "ロック解除リクエスト作成",
        "description": "年度締め済み台帳のロック解除をリクエストします",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUnlockRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "作成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/UnlockRequest" }
                  }
                }
              }
            }
          },
          "409": { "description": "既存の pending リクエストあり" }
        }
      },
      "get": {
        "tags": ["Unlock"],
        "summary": "ロック解除リクエスト一覧",
        "description": "ロック解除リクエストの一覧を取得します",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "ステータスでフィルタ",
            "schema": {
              "type": "string",
              "enum": ["pending", "approved", "rejected"]
            }
          },
          {
            "name": "ledger_id",
            "in": "query",
            "description": "台帳 ID でフィルタ",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/UnlockRequest" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/unlock-requests/{id}": {
      "get": {
        "tags": ["Unlock"],
        "summary": "ロック解除リクエスト詳細",
        "description": "ロック解除リクエストの詳細を取得します",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/UnlockRequest" }
                  }
                }
              }
            }
          },
          "404": { "description": "見つからない" }
        }
      }
    },
...
    "/api/admin/election-requests": {
      "get": {
        "tags": ["Admin"],
        "summary": "選挙リクエスト一覧",
        "description": "選挙登録リクエストの一覧を取得します",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "approved", "rejected", "needs_info"]
            }
          }
        ],
        "responses": {
          "200": { "description": "成功" }
        }
      }
    },
    "/api/admin/election-requests/{id}/approve": {
      "put": {
        "tags": ["Admin"],
        "summary": "選挙リクエスト承認",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "area_code": {
                    "type": "string",
                    "description": "正式な選挙区コード"
                  },
                  "reviewed_by": { "type": "string" }
                },
                "required": ["area_code"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "承認成功" },
          "409": { "description": "選挙 ID が既に存在" }
        }
      }
    },
    "/api/admin/election-requests/{id}/reject": {
      "put": {
        "tags": ["Admin"],
        "summary": "選挙リクエスト却下",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rejection_reason": { "type": "string" },
                  "reviewed_by": { "type": "string" }
                },
                "required": ["rejection_reason"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "却下成功" }
        }
      }
    },
    "/api/admin/organization-requests": {
      "get": {
        "tags": ["Admin"],
        "summary": "団体リクエスト一覧",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "成功" }
        }
      }
    },
    "/api/admin/organization-requests/{id}/approve": {
      "put": {
        "tags": ["Admin"],
        "summary": "団体リクエスト承認",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "politician_id": {
                    "type": "string",
                    "description": "紐づける政治家 ID"
                  },
                  "reviewed_by": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "承認成功" }
        }
      }
    },
    "/api/admin/organization-requests/{id}/reject": {
      "put": {
        "tags": ["Admin"],
        "summary": "団体リクエスト却下",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rejection_reason": { "type": "string" },
                  "reviewed_by": { "type": "string" }
                },
                "required": ["rejection_reason"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "却下成功" }
        }
      }
    },
    "/api/admin/unlock-requests": {
      "get": {
        "tags": ["Admin"],
        "summary": "ロック解除リクエスト一覧（管理者）",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "成功" }
        }
      }
    },
    "/api/admin/unlock-requests/{id}/approve": {
      "put": {
        "tags": ["Admin"],
        "summary": "ロック解除承認",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "approved_by": { "type": "string" },
                  "unlock_days": {
                    "type": "integer",
                    "default": 7,
                    "description": "解除有効日数"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "承認成功" }
        }
      }
    },
    "/api/admin/unlock-requests/{id}/reject": {
      "put": {
        "tags": ["Admin"],
        "summary": "ロック解除却下",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rejection_reason": { "type": "string" },
                  "rejected_by": { "type": "string" }
                },
                "required": ["rejection_reason"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "却下成功" }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "管理者一覧",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AdminUser" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "管理者招待",
        "description": "Magic Link で新しい管理者を招待します",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "name": { "type": "string" },
                  "role": {
                    "type": "string",
                    "enum": ["admin", "super_admin"],
                    "default": "admin"
                  }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "招待成功" },
          "409": { "description": "メールアドレスが既に登録済み" }
        }
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "管理者詳細",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "成功" },
          "404": { "description": "見つからない" }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "管理者情報更新",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "role": {
                    "type": "string",
                    "enum": ["admin", "super_admin"]
                  },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "管理者削除（論理削除）",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "削除成功" }
        }
      }
    },
    "/api/admin/users/{id}/reset-password": {
      "post": {
        "tags": ["Admin"],
        "summary": "パスワードリセットメール送信",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "送信成功" },
          "404": { "description": "管理者が見つからない" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API キー（開発用: dev_xxx, 本番用: prod_xxx）"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase Auth アクセストークン"
      }
    },
    "schemas": {
...
      "CreateUnlockRequest": {
        "type": "object",
        "properties": {
          "ledger_id": { "type": "string", "format": "uuid" },
          "ledger_type": {
            "type": "string",
            "enum": ["election", "organization"]
          },
          "fiscal_year": {
            "type": "integer",
            "description": "organization の場合は必須"
          },
          "requested_by_user_id": { "type": "string", "format": "uuid" },
          "requested_by_email": { "type": "string", "format": "email" },
          "reason": { "type": "string", "minLength": 10 }
        },
        "required": [
          "ledger_id",
          "ledger_type",
          "requested_by_user_id",
          "requested_by_email",
          "reason"
        ]
      },
...
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["admin", "super_admin"] },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}